Expand on endpoint selection strategy
This commit allows for a broader endpoint selection strategy for load
testing. Specifically, there are 3 possible strategies for selecting
endpoints for load testing:

1. "supplied" - use only supplied endpoints for load testing.
2. "discovered" - use only discovered endpoints for load testing (and
   none of the supplied ones).
3. "any" - use both supplied and discovered endpoints for testing.

This allows for permutations such as when one supplies a seed node as an
endpoint, but one does not want to submit transactions to that seed
node.

This necessarily introduces another command line parameter
`--max-endpoints`, which allows one to restrict the maximum number of
endpoints returned for load testing after applying the selection
strategy.