Ensure kvstore client honours tx size
This commit adds the necessary functionality for a key/value store
client to honour the preconfigured transaction size (previously it did
not).

It also necessarily introduces a breaking change to the client API,
which now requires a `ValidateConfig` method on `ClientFactory`
instances to allow for client-specific configuration validation. It also
now requires that the overall application configuration be passed to the
`NewClient` method on the `ClientFactory` such that, in the case of the
key/value store client, it can instantiate a client with the required
transaction size.