Add tx byte tracking and testing
This commit facilitates tracking of total transaction bytes sent across
all slaves for master/slave mode, and for the whole load test for
standalone mode. These statistics are now output via Prometheus in
master/slave mode, and in the aggregate statistics in both modes.

Integration testing is added for validation of these statistics.