Update module name to match the repository.
Since this module was moved to a new repository, the go tool complains on
import that its declared name doesn't match its requirement spec:

 module declares its path as: github.com/interchainio/tm-load-test
        but was required as: github.com/informalsystems/tm-load-test

This updates the module declaration and the internal import paths, along with
the build string injection rule in the Makefile.

Since the repository was forwarded, and all the existing issues still work at
the new address, I also updated the changelog for consistency.